refactor: to accomodate latest changes
