+++ src/components/List/ListItem/Variants/SingleSelectItem/index.ts
@@ -1,0 +1,1 @@
+export { SingleSelectItem } from './SingleSelectItem';
