--- src/components/List/ListItem/Variants/SuggestionItem/index.ts
@@ -1,1 +1,0 @@
-export { SuggestionItem, Props as SuggestionItemProps } from './SuggestionItem';
